Don't override existing .zshrc.local

+++ zsh/install.sh
@@ -4,4 +4,6 @@ HERE="$(pwd)"
 
 set -e -u -x
 
-cp $HERE/zshrc.local $HOME/.zshrc.local
+if [ ! -e "$HERE/zshrc.local"]; then
+    cp $HERE/zshrc.local $HOME/.zshrc.local
+fi
